IncusDesktop_Agent/Agent/requirements.txt
httpx>=0.27.02
pydantic
flask[async]
flask-sock
typer
rich
python-dotenv